Document Back mouse button in PointerEvent message
The highest bit in *button-mask* has never been defined to be anything,
but it looks like there are some implementations that respond to this
bit being set.

This is likely due to just passing the button mask as is directly to
X11, which interprets this bit as being the back button commonly found
on the side of mice.